Traversal code completely replaced to operate on a copy
Since the copy is returned instead of modifying the original this is an
interface breaking change.

The change was driven by discovery of a bug in unsetting slice elements.
Given a slice of N elements, unsetting 1 element should have resulted in
a slice with N-1 elements. Internally this happened but as soon as the
slice value was accessed the slice would duplicate the last element of
the slice to give it N elements again. This was true for unsetting any
number of elements e.g. unsetting 2 elements resulted in 3 copies of
what should have been the last element, unsetting 3 resulted in 4 etc.